requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.2
line_bot_sdk==3.9.0
plurk_oauth==0.9.2
python-dotenv==1.0.1
Requests==2.31.0
google-genai==0.3.0
